requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
python-dotenv==1.0.1
stripe==10.12.0
httpx==0.27.2
Pillow==10.4.0
aiofiles==24.1.0
email-validator==2.2.0
cryptography==41.0.7
phonenumbers==8.13.50
APScheduler==3.10.4
pytest==8.3.3
openai==2.18.0
PyJWT[crypto]==2.9.0
slowapi==0.1.9
